+++ src/VM/VM.cpp
@@ -128,17 +128,9 @@ namespace Ark
                         ++m_ip;
                     }
                 }
-                catch (const Ark::TypeError& e)
+                catch (const std::exception& e)
                 {
                     Ark::logger.error(e.what());
-                }
-                catch (const Ark::ZeroDivisionError& e)
-                {
-                    Ark::logger.error(e.what());
-                }
-                catch (const std::runtime_error& e)
-                {
-                    Ark::logger.error("RuntimeError", e.what());
                 }
             }
         }